refactor arena configs
